Update README.md
Sweet Home is frantic, it's scary, pushing you to the edge of your seat, and the next moment you're crying like a baby. The next moment seems to be always a surprise episode after episode, and to me, that's what a good story is like; The Ring, the acting, the atmosphere, the visuals, the twist ending. It was a modernized update on “chain letters” where you receive a creepy letter; And last Hereditary, because its scary and I love scary stuff.